Rearrange some more code to move swagger specifics out of types
